+++ tests/ui/indexing_slicing.rs
